requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.4
django-jazzmin==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
pdf2image==1.17.0
pillow==11.1.0
PyMySQL==1.1.1
PyPDF2==3.0.1
reportlab==4.3.1
sqlparse==0.5.3
tzdata==2024.2
